+++ client/components/waiting-room.js
@@ -4,57 +4,12 @@ import { withRouter } from 'react-router-dom'
 import firebase from '../firebase'
 
 import '../css/_room.scss';
+import { divvySpaces } from './divvySpaces';
 
 const WaitingRoom = (props) => {
   const { board, user } = props
   const maxPlayers = board.maxPlayers
   const numPlayers = board.state.playerOrder.length
-  const players = ['null', ...board.state.playerOrder]
-
-  const divvySpaces = (evnt) => {
-    evnt.preventDefault();
-    let numPlayerSpaces;
-    let numSpaces = Object.keys(board.hexes).length
-    let numVoidSpaces = Math.floor(numSpaces / 10) * 2;
-    let numAllotSpaces = numSpaces - numVoidSpaces;
-
-    if (numPlayers >= 1) {
-      if (numAllotSpaces % numPlayers !== 0) {
-        let numExtra = numAllotSpaces % numPlayers;
-        numVoidSpaces += numExtra;
-        numAllotSpaces -= numExtra;
-      }
-
-      numPlayerSpaces = numAllotSpaces / numPlayers;
-
-      let numRed = numPlayerSpaces;
-      let numOrange = numPlayerSpaces;
-      let numYellow = numPlayerSpaces;
-      let numGreen = numPlayerSpaces;
-      let numBlue = numPlayerSpaces;
-
-      let assignmentColors = [
-        {color: 'black', amount: numVoidSpaces},
-        {color: 'red', amount: numRed},
-        {color: 'orange', amount: numOrange},
-        {color: 'yellow', amount: numYellow},
-        {color: 'green', amount: numGreen},
-        {color: 'blue', amount: numBlue}]
-
-      Object.keys(board.hexes).forEach(id => {
-        while (board.hexes[id].playerId === '') {
-          let assign = Math.floor(Math.random() * (numPlayers + 1));
-          if (assignmentColors[assign].amount) {
-            assignmentColors[assign].amount--
-            board.hexes[id].playerId = players[assign]
-            firebase.ref(`/boards/${props.match.params.boardId}/hexes/${id}`).update({playerId: players[assign]})
-          }
-        }
-      })
-    }
-
-    props.startGame()
-  }
 
   return (
     <div id="waiting-room">
@@ -68,7 +23,7 @@ const WaitingRoom = (props) => {
               <button
                 className="text"
                 type="submit"
-                onClick={divvySpaces}
+                onClick={(evnt) => props.startGame(evnt, board.state.playerOrder, board.hexes)}
               >
                 Start Game
               </button>
@@ -86,9 +41,11 @@ const mapState = state => {
 
 const mapDispatch = (dispatch, ownProps) => {
   return {
-    startGame(){
-      const boardId = ownProps.match.params.boardId
-      firebase.ref(`/boards/${boardId}/state`).update({status: 'playing'})
+    startGame(evnt, playerOrder, hexes){
+      evnt.preventDefault();
+      const boardId = ownProps.match.params.boardId;
+      divvySpaces(playerOrder, hexes, boardId);
+      firebase.ref(`/boards/${boardId}/state`).update({status: 'playing'});
     }
   }
 }
